moved ApiReader call back to app.js
