+++ movie-browser/src/App.js
@@ -1,15 +1,33 @@
 /* eslint-disable no-unused-vars */
 import React from 'react'
-import { CardContainer } from './CardContainer'
+import CardContainer from './CardContainer'
+import ApiReader from './model/ApiReader';
 
-export const url = 'https://ghibliapi.herokuapp.com/films'
+// *** constant(s)
+const url = 'https://ghibliapi.herokuapp.com/films'
 
-const App = () => {
-  return (
-    <div>
-      <CardContainer />
-    </div>
-  )
+class App extends React.Component {
+  state = {movies: []};
+
+  // *** running ApiReader when mounted
+  componentDidMount() {
+    this.fetchData();
+  }
+
+  async fetchData() {
+    const data = await ApiReader(url);
+    this.setState({ movies: data });
+  }
+
+  // *** display
+  render() {
+    const { movies } = this.state;
+    return (
+      <div>
+        <CardContainer movies={movies}  />
+      </div>
+    )
+  }
 }
 
 export default App
