+++ movie-browser/src/CardContainer.js
@@ -1,28 +1,15 @@
-import React from 'react';
-import Card from './view/Card';
-import ApiReader from './model/ApiReader';
-import { url } from './App';
+/* eslint-disable no-unused-vars */
+import React from 'react'
+import Card from './view/Card'
 
-export class CardContainer extends React.Component {
-  state = {movies: []};
-
-  componentDidMount() {
-    this.fetchData();
-  }
-
-  async fetchData() {
-    const data = await ApiReader(url);
-    this.setState({ movies: data });
-  }
-
-  render() {
-    const { movies } = this.state;
-    return (
-        <div className='row'>
-            {this.state.movies.map(movie => 
-                <Card key={movie.id} {...movie} />
-            )}
-        </div>
-    )
-  }
+const CardContainer = (props) => {
+  return (
+    <div className='row'>
+      {props.movies.map(movie =>
+        <Card key={movie.id} {...movie} />
+      )}
+    </div>
+  )
 }
+
+export default CardContainer
